requirements.txt
certifi==2020.4.5.1
numpy==1.18.3
opencv-python==4.2.0.34
wincertstore==0.2
flask==1.1.2
requests-toolbelt==0.9.1
pillow==7.1.2